+++ app/models/concerns/localizable.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
+# TODO: Remove in v17.0.0
 module Localizable
   extend ActiveSupport::Concern
 
   module ClassMethods
     def localize(attr)
-
       localized_attr = [attr, I18n.locale].join('_')
 
       unless column_names.include? localized_attr
@@ -16,7 +16,5 @@ module Localizable
         self[localized_attr]
       end
     end
-
   end
-
 end