+++ app/models/concerns/random_name_generator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Generates random forum name
+# Generates random forum name.
 module RandomNameGenerator
   extend ActiveSupport::Concern
 
@@ -8,7 +8,7 @@ module RandomNameGenerator
     def self.generate_forum_name(input, additional_seed = nil)
       input += additional_seed if additional_seed
       seed = Digest::MD5.hexdigest(input.to_s).hex.to_s
-      adjective = adjectives[(seed[0..3].to_i % adjectives.size)]
+      # adjective = adjectives[(seed[0..3].to_i % adjectives.size)]
       color = colors[(seed[4..7].to_i % colors.size)]
       animal = animals[(seed[8..11].to_i % animals.size)]
       # "#{adjective}#{color}#{animal}#{seed[12..15]}"