+++ app/models/topic_user.rb
@@ -2,11 +2,11 @@
 
 # Tracks how far has read through a forum topic.
 class TopicUser < ApplicationRecord
-  # Model Validation
+  # Validations
   validates :topic_id, :user_id, presence: true
   validates :user_id, uniqueness: { scope: :topic_id }
 
-  # Model Relationships
+  # Relationships
   belongs_to :topic
   belongs_to :user
 end